Post Dynamics rewrite + Basic Dark Light theming

+++ app/src/lib/server/services/imageService.ts
@@ -1,22 +1,36 @@
 import { ImageModel } from "../models/imageModel"
 import { type AppImageData } from "../models/imageModel"
 import { FileModel } from "../models/fileModel"
-import { imageFileUtil } from "../utility/imageUtil"
 
+import sharp from "sharp";
 import 'fluent-ffmpeg';
 import Ffmpeg from "fluent-ffmpeg";
 
 
 export const imageService = {
+    async createImageThumbnail(imgBuffer: Uint8Array , scale: number = 200): Promise<Buffer> { // want to add webm support for thumbnail creation TODO
+        // file extension check here
+        try {
+            return await sharp(imgBuffer)
+                .resize({ width: scale})
+                .webp()
+                .toBuffer();
+        } catch (error) {
+            throw new Error(`Error creating thumbnail: ${error}`);
+        }
+    },
+
     async updateThumbnail(image: AppImageData, buffer?: Buffer , overwrite = false){
 
-        const thumbnailExists = await FileModel.checkExists(image.thumbnailPath);
+        if (image.thumbnailPath) {
+            const thumbnailExists = await FileModel.checkExists(image.thumbnailPath);
 
-        if (thumbnailExists) {
-            if (overwrite) {
-                await FileModel.delete(image.thumbnailPath);
-            } else {
-                return false;
+            if (thumbnailExists) {
+                if (overwrite) {
+                    await FileModel.delete(image.thumbnailPath);
+                } else {
+                    return false;
+                }
             }
         }
 
@@ -25,7 +39,7 @@ export const imageService = {
     
         try{
             buffer = buffer ? buffer : await FileModel.read(image.imagePath);
-            const thumbnail = await imageFileUtil.createImageThumbnail(buffer);
+            const thumbnail = await this.createImageThumbnail(buffer);
             
             await ImageModel.updateImage(image._id, "thumbnailPath", thumbnailPath);
 
@@ -45,17 +59,19 @@ export const imageService = {
 
     async updateVideoThumbnail(imageData: AppImageData, overwrite = false){
 
-        const thumbnailExists = await FileModel.checkExists(imageData.thumbnailPath);
-        const thumbnailname = `${imageData.uploadDate}_thmb.webp`;
-        const thumbnailPath = `thumb/${thumbnailname}`;
-
-        if (thumbnailExists) {
-            if (overwrite) {
-                await FileModel.delete(imageData.thumbnailPath);
-            } else {
-                return false;
+        if (imageData.thumbnailPath) {
+            const thumbnailExists = await FileModel.checkExists(imageData.thumbnailPath);
+            if (thumbnailExists) {
+                if (overwrite) {
+                    await FileModel.delete(imageData.thumbnailPath);
+                } else {
+                    return false;
+                }
             }
         }
+        
+        const thumbnailname = `${imageData.uploadDate}_thmb.webp`;
+        const thumbnailPath = `thumb/${thumbnailname}`;
 
         try{
             await new Promise((resolve, reject) => {
